+++ familyhelper-life-impl/src/test/java/com/dwarfeng/familyhelper/life/impl/configuration/ServiceExceptionMapperConfiguration.java
@@ -48,6 +48,8 @@ public class ServiceExceptionMapperConfiguration {
         destination.put(ActivityLockedException.class, ServiceExceptionCodes.ACTIVITY_LOCKED);
         destination.put(ActivityCoverNotExistsException.class, ServiceExceptionCodes.ACTIVITY_COVER_NOT_EXISTS);
         destination.put(IllegalActivityCoverStateException.class, ServiceExceptionCodes.ILLEGAL_ACTIVITY_COVER_STATE);
+        destination.put(ActivityFileNotExistsException.class, ServiceExceptionCodes.ACTIVITY_FILE_NOT_EXISTS);
+        destination.put(IllegalActivityFileStateException.class, ServiceExceptionCodes.ILLEGAL_ACTIVITY_FILE_STATE);
         return new MapServiceExceptionMapper(destination, com.dwarfeng.subgrade.sdk.exception.ServiceExceptionCodes.UNDEFINE);
     }
 }